+++ src/Molecules/LabeledValue/LabeledValue.types.ts
@@ -1,8 +1,9 @@
 import React, { ReactNode } from 'react';
 
 export type Props = {
-  label: string | React.ReactNode;
   children: ReactNode;
+  className?: string;
+  label: string | React.ReactNode;
 };
 
 export type LabeledValueComponent = React.FunctionComponent<Props>;